+++ structurizr-core/src/com/structurizr/model/DeploymentElement.java
@@ -1,7 +1,7 @@
 package com.structurizr.model;
 
 /**
- * This is the superclass for model elements that describe deployment nodes and container instances.
+ * This is the superclass for model elements that describe deployment nodes, infrastructure nodes, and container instances.
  */
 public abstract class DeploymentElement extends Element {