+++ structurizr-core/src/com/structurizr/model/Tags.java
@@ -11,6 +11,7 @@ public class Tags {
     public static final String COMPONENT = "Component";
 
     public static final String DEPLOYMENT_NODE = "Deployment Node";
+    public static final String INFRASTRUCTURE_NODE = "Infrastructure Node";
     public static final String CONTAINER_INSTANCE = "Container Instance";
 
     /**